moved component folder into app directory
