+++ app/components/Hero.jsx
