+++ app/layout.jsx
@@ -1,8 +1,8 @@
 import "../styles/globals.css";
 import { Inter } from "next/font/google";
 
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
+import Navbar from "./components/Navbar";
+import Footer from "./components/Footer";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
