+++ app/page.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Hero from "@/components/Hero";
+import Hero from "@/app/components/Hero";
 
 export default function Home() {
   return (
